refactore controller without service
