+++ backend/config/services.yaml
@@ -7,6 +7,7 @@ parameters:
     upload_directory: '%kernel.project_dir%/public/uploads'
 
 services:
+
     # default configuration for services in *this* file
     _defaults:
         autowire: true      # Automatically injects dependencies in your services.
