+++ backend/src/Controller/ProjectController.php
@@ -2,16 +2,38 @@
 
 namespace App\Controller;
 
+use App\Entity\Projects;
+use App\Entity\ProjectsImages;
+use App\Repository\PortfoliosRepository;
+use App\Repository\ProjectsRepository;
+use App\Service\FileUploaderService;
 use App\Service\ProjectService;
+use App\Service\ValidatorBaseService;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Request;
 use OpenApi\Annotations as OA;
+use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class ProjectController extends AbstractController
 {
+    public function __construct(
+        private EntityManagerInterface $entityManager,
+        private ProjectsRepository     $projectsRepository,
+        private SerializerInterface    $serializer,
+        private PortfoliosRepository   $portfoliosRepository,
+        private ValidatorInterface     $validator,
+        private FileUploaderService    $fileUploader,
+        private ValidatorBaseService   $validatorBaseService
+    )
+    {
+    }
+
     /**
      * @OA\Post(
      *     path="/api/project",
@@ -43,8 +65,8 @@ class ProjectController extends AbstractController
      */
     #[Route('/api/project', methods: ['POST'])]
     public function add_project(
-        Request             $request,
-        ProjectService      $projectService,
+        Request        $request,
+        ProjectService $projectService,
     ): JsonResponse
     {
         $this->denyAccessUnlessGranted('ROLE_USER');
@@ -61,7 +83,34 @@ class ProjectController extends AbstractController
             return new JsonResponse(['error' => 'bad request'], Response::HTTP_BAD_REQUEST);
         }
 
-        $jsonProject = $projectService->createProject($user, $data, $files, $uploadDir);
+        $portfolio = $this->portfoliosRepository->findOneBy(['users' => $user]);
+
+        if (!$portfolio) {
+            return new JsonResponse(['error' => 'no portfolio found'], Response::HTTP_NOT_FOUND);
+        }
+
+        $project = $this->serializer->deserialize($data, Projects::class, 'json');
+        $project->setPortfolio($portfolio);
+
+        $errors = $this->validator->validate($project);
+        $this->validatorBaseService->catchInvalidData($errors);
+
+        foreach ($files as $file) {
+            if (!$file instanceof UploadedFile) {
+                throw new \InvalidArgumentException("Invalid file format.");
+            }
+
+            $filePath = $this->fileUploader->uploadFile($file, $uploadDir);
+            $projectImage = new ProjectsImages();
+            $projectImage->setImgSrc($filePath);
+            $projectImage->setImgAlt($project->getTitle());
+            $project->addProjectsImage($projectImage);
+        }
+
+        $this->entityManager->persist($project);
+        $this->entityManager->flush();
+
+        $jsonProject = $this->serializer->serialize($project, 'json', ['groups' => 'getPortfolio']);
         return new JsonResponse($jsonProject, Response::HTTP_CREATED, [], true);
     }
 
@@ -104,7 +153,6 @@ class ProjectController extends AbstractController
     public function update_project(
         string         $id,
         Request        $request,
-        ProjectService $projectService
     ): JsonResponse
     {
         $this->denyAccessUnlessGranted('ROLE_USER');
@@ -119,7 +167,19 @@ class ProjectController extends AbstractController
             return new JsonResponse(['error' => 'bad request'], Response::HTTP_BAD_REQUEST);
         }
 
-        $jsonProject = $projectService->UpdateProject($user, $data, $id);
+        $project = $this->projectsRepository->findOneBy(['id' => $id]);
+
+        if ($project->getPortfolio() !== $this->portfoliosRepository->findOneBy(['users' => $user])) {
+            return new JsonResponse(['error' => " no project found "], Response::HTTP_NOT_FOUND);
+        }
+
+        $this->serializer->deserialize($data, Projects::class, 'json', ['object_to_populate' => $project]);
+        $errors = $this->validator->validate($project);
+        $this->validatorBaseService->catchInvalidData($errors);
+
+        $this->entityManager->flush();
+
+        $jsonProject = $this->serializer->serialize($project, 'json', ['groups' => 'getPortfolio']);
         return new JsonResponse($jsonProject, Response::HTTP_OK, [], true);
     }
 }
