+++ backend/src/Service/ValidatorBaseService.php
@@ -3,9 +3,9 @@ namespace App\Service;
 
 use Symfony\Component\Validator\ConstraintViolationListInterface;
 
-abstract class ValidatorBaseService
+class ValidatorBaseService
 {
-    protected function CatchInvalidData(ConstraintViolationListInterface $errors): void
+    public function CatchInvalidData(ConstraintViolationListInterface $errors): void
     {
         if (count($errors) > 0) {
             $errorMessages = [];
